requirements.txt
requests>=2.28
pyodbc
azure-identity
azure-functions>=1.11.0
azure-functions-durable
pandas
pandas_ta
python-binance
azure-communication-email
